requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Qwen LLM API
dashscope>=1.20.0
openai>=1.50.0

# WebSocket support
websockets>=14.0

# Database
aiosqlite>=0.20.0
SQLAlchemy>=2.0.36

# Authentication
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.17

# HTTP client
httpx>=0.27.0

# CLI interface
click>=8.1.7
rich>=13.9.0

# Utilities
python-dateutil>=2.9.0
